Establish framework to display characters
Establish a framework to display characters on the display on command. A framebuffer is created storing the current state of the display, and it is updated once/loop, iterating through the different sections to be addressed. A function is added to iterate through the printable ascii and display each char in sequence on the display.