Added game modes, button code
Big commit time, this commit adds a main game mode that quizzes the user on all ascii characters, a 'table' mode that allows the user to browse a virtual ascii table, and a mode wherein you can put in any binary number and it'll give you the corresponding character. also added code to poll the buttons and the nibble switch.